# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Compass - Student Lifecycle Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EnergyLog {
  id: number;
  taskName: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  energyLevel: number;
  categoryId?: number | null;
  categoryName?: string | null;
  notes?: string | null;
  loggedAt: string;
  createdAt: string;
}

export interface CreateEnergyLogBody {
  taskName: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  energyLevel: number;
  categoryId?: number | null;
  notes?: string | null;
  loggedAt?: string | null;
}

export interface UpdateEnergyLogBody {
  taskName?: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  energyLevel?: number;
  categoryId?: number | null;
  notes?: string | null;
}

export interface Category {
  id: number;
  name: string;
  logCount: number;
  avgEnergy?: number | null;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
}

export interface UpdateCategoryBody {
  name: string;
}

export interface Mentor {
  id: number;
  name: string;
  major: string;
  year: string;
  bio: string;
  avatarUrl?: string | null;
  strengths: string[];
  rating?: number | null;
  sessionCount: number;
  isAvailable: boolean;
  createdAt: string;
}

export interface CreateMentorBody {
  name: string;
  major: string;
  year: string;
  bio: string;
  avatarUrl?: string | null;
  strengths: string[];
  isAvailable?: boolean;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Session {
  id: number;
  mentorId: number;
  mentorName: string;
  mentorMajor: string;
  mentorAvatarUrl?: string | null;
  topic: string;
  notes?: string | null;
  status: SessionStatus;
  scheduledAt: string;
  createdAt: string;
}

export interface CreateSessionBody {
  mentorId: number;
  topic: string;
  notes?: string | null;
  scheduledAt: string;
}

export type UpdateSessionBodyStatus =
  (typeof UpdateSessionBodyStatus)[keyof typeof UpdateSessionBodyStatus];

export const UpdateSessionBodyStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateSessionBody {
  topic?: string;
  notes?: string | null;
  status?: UpdateSessionBodyStatus;
  scheduledAt?: string;
}

export interface MajorRecommendation {
  id: number;
  majorName: string;
  matchScore: number;
  reasoning: string;
  topCategories: string[];
  careerPaths: string[];
  createdAt: string;
}

export interface DashboardSummary {
  totalLogs: number;
  avgEnergyOverall?: number | null;
  topCategory?: string | null;
  topCategoryAvgEnergy?: number | null;
  burnoutRiskCategories: number;
  thrivingCategories: number;
  upcomingSessions: number;
  totalMentors: number;
  streak: number;
}

export type EnergyByCategoryStatus =
  (typeof EnergyByCategoryStatus)[keyof typeof EnergyByCategoryStatus];

export const EnergyByCategoryStatus = {
  thriving: "thriving",
  okay: "okay",
  "needs-attention": "needs-attention",
} as const;

export interface EnergyByCategory {
  categoryId: number;
  categoryName: string;
  avgEnergy: number;
  logCount: number;
  status: EnergyByCategoryStatus;
}

export type GetEnergyLogsParams = {
  categoryId?: number;
  limit?: number;
};

export type GetMentorsParams = {
  major?: string;
  year?: string;
};